Better settings to control manual learning.
see #622

+++ modoboa/extensions/amavis/general_callbacks.py
@@ -8,6 +8,8 @@ from modoboa.extensions.amavis.lib import (
     create_user_and_policy, update_user_and_policy, delete_user_and_policy,
     create_user_and_use_policy, delete_user
 )
+from .lib import manual_learning_enabled
+from .models import Users
 
 
 @events.observe("UserMenuDisplay")
@@ -54,16 +56,51 @@ def on_domain_alias_deleted(domainaliases):
         delete_user("@{0}".format(domainalias.name))
 
 
-@events.observe("MailboxCreated")
-def on_mailbox_created(user, mailbox):
-    """Create amavis records for the new mailbox."""
-    create_user_and_policy(mailbox.full_address)
+@events.observe("MailboxModified")
+def on_mailbox_modified(mailbox):
+    """Update amavis records if address has changed."""
+    if parameters.get_admin("MANUAL_LEARNING") == "no" or \
+       mailbox.full_address == mailbox.old_full_address:
+        return
+    user = Users.objects.select_related.get(email=mailbox.old_full_address)
+    full_address = mailbox.full_address
+    user.email = full_address
+    user.policy.policy_name = full_address[:32]
+    user.policy.sa_username = full_address
+    user.policy.save()
+    user.save()
 
 
 @events.observe("MailboxDeleted")
 def on_mailbox_deleted(mailbox):
-    """Delete amavis records related to this mailbox."""
+    """Clean amavis database when a mailbox is removed."""
+    if parameters.get_admin("MANUAL_LEARNING") == "no":
+        return
     delete_user_and_policy(mailbox.full_address)
+
+
+@events.observe("MailboxAliasCreated")
+def on_mailboxalias_created(user, alias):
+    """Create amavis record for the new alias.
+
+    FIXME: how to deal with distibution lists ?
+    """
+    if not manual_learning_enabled(user) or alias.type != "alias":
+        return
+    policy = Policy.objects.get(
+        policy_name=alias.mboxes.all()[0].full_address
+    )
+    email = alias.full_address
+    Users.objects.create(email=email, policy=policy, fullname=email)
+
+
+@events.observe("MailboxAliasDeleted")
+def on_mailboxalias_deleted(alias):
+    """Clean amavis database when an alias is removed."""
+    if parameters.get_admin("MANUAL_LEARNING") == "no":
+        return
+    if Users.objects.exists(email=alias.fullname):
+        Users.objects.delete(email=alias.fullname)
 
 
 @events.observe("GetStaticContent")
